legal_agentic_system/requirements.txt
Flask
Werkzeug
huggingface_hub
pdf2image
Pillow
langchain
langchain-groq
langchain-community
langchain-huggingface
langchain-text-splitters
langgraph
tavily-python
faiss-cpu
python-dotenv
pypdf
sentence-transformer